+++ src/utils/test/setup.js
@@ -1,0 +1,11 @@
+import "@testing-library/jest-dom";
+
+// const mockedUseNavigate = vi.fn();
+
+// vi.mock("react-router-dom", async () => {
+//   const actual = await vi.importActual("react-router-dom");
+//   return {
+//     ...actual,
+//     useNavigate: () => mockedUseNavigate,
+//   };
+// });