+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Version History
 
+### [v0.3.0] - `2023-06-16`
+
+- Added vitest for testing tool
+- Updated "All-in-one" files and structures
+- Updated dependencies
+
 ### [v0.2.1] - `2023-05-22`
 
 - Updated dependencies
@@ -16,6 +22,7 @@
 
 - First time ever, it contains a lot of dir/files but it was needed.
 
+[v0.3.0]: https://github.com/buildingwatsize/vite-react-tailwind-antd/releases/tag/v0.3.0
 [v0.2.1]: https://github.com/buildingwatsize/vite-react-tailwind-antd/releases/tag/v0.2.1
 [v0.2.0]: https://github.com/buildingwatsize/vite-react-tailwind-antd/releases/tag/v0.2.0
 [v0.1.0]: https://github.com/buildingwatsize/vite-react-tailwind-antd/releases/tag/v0.1.0